Add files via upload
# Problem Statement

In this notebook we will be discussing about the Credit Card Fraud Classification and will be building a Supervised Learning Model to predict the Fraud Transaction that may have happened. Till this date Fraud Detection is one of the major problems of the banks and companies. Solving this problem will help the banks to identify the Fraud Transaction that may have happened and it will help them to take actions against that. This will reduce the risk of the Fraud Transaction. It will also reduce a lot of manual power that till is used to identify the Fraud Detection and will increase the productivity of the Bank. In India in the month of September about 81 Crore of Fraud Detection has been spotted.